requirements.txt
# Core Logic
pytesseract==0.3.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
Pillow==10.0.0
numpy==1.26.0

# CLI and Utilities
typer[all]==0.9.0
rich==13.6.0
python-dotenv==1.0.0